+++ pyodide-build/pyodide_build/tests/test_cli.py
@@ -134,29 +134,6 @@ def test_config_get(cfg_name, env_var):
     assert result.stdout.strip() == common.get_make_flag(env_var)
 
 
-def test_fetch_or_build_pypi(selenium, tmp_path):
-    # TODO: Run this test without building Pyodide
-
-    output_dir = tmp_path / "dist"
-    # one pure-python package (doesn't need building) and one sdist package (needs building)
-    pkgs = ["pytest-pyodide", "pycryptodome==3.15.0"]
-
-    os.chdir(tmp_path)
-
-    app = typer.Typer()
-    app.command()(build.main)
-
-    for p in pkgs:
-        result = runner.invoke(
-            app,
-            [p],
-        )
-        assert result.exit_code == 0, result.stdout
-
-    built_wheels = set(output_dir.glob("*.whl"))
-    assert len(built_wheels) == len(pkgs)
-
-
 def test_create_zipfile(temp_python_lib, tmp_path):
     from zipfile import ZipFile