+++ src/tests/test_cmdline_runner.py
@@ -156,16 +156,14 @@ def venv(runtime):
 @needs_emscripten
 def test_venv_success_log(selenium, capsys):
     with venv_ctxmgr(Path(".venv-pyodide-tmp-test")):
-        msg = dedent(
-            """
-            Creating Pyodide virtualenv at .venv-pyodide-tmp-test
-            ... Configuring virtualenv
-            ... Installing standard library
-            Successfully created Pyodide virtual environment!
-        """
-        ).strip()
+        msgs = [
+            "Creating Pyodide virtualenv at .venv-pyodide-tmp-test",
+            "... Configuring virtualenv",
+            "... Installing standard library",
+            "Successfully created Pyodide virtual environment!",
+        ]
     captured = capsys.readouterr()
-    assert captured.out.strip() == msg
+    assert [l.strip() for l in captured.out.splitlines()] == msgs
     assert captured.err == ""
 
 
@@ -184,8 +182,7 @@ def test_venv_fail_log(selenium, capsys):
     captured = capsys.readouterr()
     assert captured.out.strip() == msg
     assert (
-        captured.err
-        == "ERROR: dest directory '.venv-pyodide-tmp-test' already exists\n"
+        "ERROR: dest directory '.venv-pyodide-tmp-test' already exists" in captured.err
     )